config: Separate [remote] and [local] sections
Separate [images] sections to [remote] and [local] sections. This will
make it easier to make the local service configurable, and it also helps
to create nicer options names like "local.socket". This structure also
correponds better with real structure of the application.

Besides that, this will allow disabling the local service when using
on engine as a proxy.

This patch is base on Nir's patch
https://gerrit.ovirt.org/c/93016/

Change-Id: I5fbf36a7478d603872112a0117836cdf5e240a1e
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>