config: Eliminate pki_dir
Replace pki_dir config, assuming /etc/pki/vdsm/ directory layout with
separate ca_file, cert_file, and key_file options. We still depend on
vdsm pki, but the dependency is only in the default configuration.

With this change the pki module is not needed, remove it.

This patch is base on Nir's patch
https://gerrit.ovirt.org/c/93018

Change-Id: I5df0bc76183b9764838d24fbc8c9c532807f3eec
Signed-off-by: Vojtech Juranek <vjuranek@redhat.com>