ssl: Create working test pki
Replace the vdsm certs with new certs for 127.0.0.1 valid for 10 years,
so we never have to generate them again. With the new certificates, we
can enable certificates verification in the tests.

The key to make this work with python 3.7 is to have a subjectAltName
with IP address[1] . Without this verifying server certificates in
the tests fail with this error:

    ssl.SSLCertVerificationError: [SSL: CERTIFICATE_VERIFY_FAILED]
    certificate verify failed: IP address mismatch, certificate is not
    valid for '127.0.0.1'. (_ssl.c:1076)

The certificates and keys were generated by the new test/genpki script.
We can run it again to generate new keys and certificates.

[1] https://bugs.python.org/issue34440

Change-Id: I7c0d9426df282c1a56d96878f4c7a81c0c60285b
Signed-off-by: Nir Soffer <nsoffer@redhat.com>