daemon: Optimize transferred calculation
When uploading or downloading big images using tiny chunks, the
operations list becomes huge, taking lot of memory and consuming
significant cpu time.

Optimize the calculation by removing completed operations from the
ticket, and keeping a sorted list of transferred ranges. When an
operation is removed, the operation range is merged into the completed
ranges list, which should be small in typical download or upload flows.

When calculating total transferred bytes, we create list of ranges from
the ongoing operations, which should be small, and merge the completed
and ongoing ranges.

Change-Id: Ib33967d9e858353037542eedbb3c68d350bf3ad4
Bug-Url: https://bugzilla.redhat.com/1592847
Signed-off-by: Nir Soffer <nsoffer@redhat.com>