tox: Add bench env and update python version
Separate "test" and "bench" environments, update python version for
Fedora rawhide (3.7), and remove python 3.5.

To run all tests and benchmarks, use:

    tox

To only the tests, use:

    tox -e "test-{py27,py36}"

To only the benchmarks, use:

    tox -e "test-{py27,py36}"

To get a list of envs, use:

    $ tox -l
    flake8
    test-py27
    test-py36
    test-py37
    bench-py27
    bench-py36
    bench-py37

The Makefile was update to use the new syntax, since we don't want to
run the benchmarks in the CI.

Similar changes applied to common tox.ini. Since we don't have any
benchmarks yet, keep the default py27,py36,py37 targets.

Change-Id: Ib44b8dd61b6ad60e3b013f5a102d0addad59cc65
Signed-off-by: Nir Soffer <nsoffer@redhat.com>