logging: Remove path=/path/to/image from images logs
We can have 5000 requests for same transfer, so it does not make sense
to log the image path for every request. We log the path for every
ticket when adding a ticket, and there is no need to repeat this info
again.

Because oVirt image paths are very long, this reduces the amount of data
for single 6G virt-v2v import by half:

version        size
-------------------
with path      677K
no path        331K

Change-Id: I2c81084bce78e546f7d03d5d8eb9ccdd45926288
Bug-Url: https://bugzilla.redhat.com/1614202
Signed-off-by: Nir Soffer <nsoffer@redhat.com>