daemon: Extend ticket on every request
We have seen too many failures to extend tickets recently. Typically
a single extend request is missing, and the upload fail one minute after
the missing extend request.

The current extend logic is complex and fragile. If engine or vdsm are
down or stuck for some reason when an extend message should be sent,
active transfers will expire, even if the client and daemon are happily
transferring data.

Change the logic so a ticket is extended after every client request. A
ticket will expire only if a client failed to do one request during
ticket timeout period.

With this change, you can kill both engine and vdsm once a transfer has
started. They have to be UP only when the client is trying to finalize
the transfer.

Change-Id: I78cd266e915aa509ba89d4b191bc0733f74d6d46
Bug-Url: https://bugzilla.redhat.com/1619019
Signed-off-by: Nir Soffer <nsoffer@redhat.com>