packaging: setup: Use IMAGEIO_PROXY_PORT for vdc option
Instead of DEFAULT_IMAGEIO_PROXY_PORT .

Without this patch, if you run engine-setup with a custom answer file
that has e.g.:

OVESETUP_CONFIG/imageioProxyPort=int:12121

Then the port '12121' will be used elsewhere (in the config etc.), but
the vdc option will be set to use the default port, 54323.

This is not a critical bug, because imageioProxyPort is not written to
the answer file by engine-setup, so stumbling upon this bug requires
manually editing the answer file.

Change-Id: I67eca2ce73af306480e93316dbbb868290f78d7b
Signed-off-by: Yedidyah Bar David <didi@redhat.com>