version: Bump to 1.5.0
We should bump master version when we start to work on a release, not
when we end the work. This make it easier to detect if we run code from
master or stable branch.

For next release I want to get rid of the WSGI server, which is a pretty
big rewrite, so it is time to bump the minor version. If we need to fix
bugs in current stable version we can release 1.4.x from that branch.

Change-Id: I4e519523e8e736eff944b1c51fc099c866386af3
Signed-off-by: Nir Soffer <nsoffer@redhat.com>