http: Use shorter maximum request line
BaseHTTPServer supports request-line of 65536 characters. This size is
an arbitrary value that a server choose to support. For Python builtin
server it makes sense to use large value, since some application may
like to support very long request URI.

In imageio we never need large URI, so there is no reason to support
such a large value. Shorten to 4096 characters, and replace the magic
number with a class constant.

Change-Id: I69caac2e5df0f86373279e4c276bd75393842f00
Signed-off-by: Nir Soffer <nsoffer@redhat.com>