http: Add constants for HTTP status codes
Replace usage of literal status code numbers with constants. This is not
very helpful with the well known codes such as 200 and 400, but nice
with the less known status codes such as 414 (Request-URI too long).

The constants are available http.client (httplib in python 2) but
including http_client in the http module is stupid, so we define our own
constants in the http module.

Change-Id: I5eaf52a7689050c1ba04ed61273a8f562584a379
Signed-off-by: Nir Soffer <nsoffer@redhat.com>