http: Optimize sending small response
If the payload is small, copy it to the header buffer, and send both in
one syscall.

Also remove the flush in the first write; If Response.wfile is
unbuffered (default) it does nothing, and it is buffered it will be
flushed on the end of the request.

This change improves throughput in 16% and 2.79X times faster compared
with current WSGI server. The difference in real application will be
much smaller, but this may be significant for zero requests or small
write requests.

With this change:

$ ./wrk -t1 -c1 --latency http://localhost:8000/bench/1
Running 10s test @ http://localhost:8000/bench/1
  1 threads and 1 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    82.27us   29.49us   2.44ms   98.29%
    Req/Sec    12.12k   791.92    13.03k    69.31%
  Latency Distribution
     50%   77.00us
     75%   85.00us
     90%   86.00us
     99%  128.00us
  121821 requests in 10.10s, 13.13MB read
Requests/sec:  12062.23
Transfer/sec:      1.30MB

Without this change:

$ ./wrk -t1 -c1 --latency http://localhost:8000/bench/1
Running 10s test @ http://localhost:8000/bench/1
  1 threads and 1 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    96.41us   36.98us   2.70ms   98.17%
    Req/Sec    10.40k   766.27    11.10k    88.12%
  Latency Distribution
     50%   89.00us
     75%   96.00us
     90%  111.00us
     99%  160.00us
  104480 requests in 10.10s, 11.26MB read
Requests/sec:  10344.61
Transfer/sec:      1.11MB

With currnet WSGI server:

$ ./wrk -t1 -c1 --latency http://localhost:8000/bench/1
Running 10s test @ http://localhost:8000/bench/1
  1 threads and 1 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   229.14us   39.60us   1.94ms   97.34%
    Req/Sec     4.38k   258.06     4.59k    82.00%
  Latency Distribution
     50%  219.00us
     75%  230.00us
     90%  247.00us
     99%  338.00us
  43587 requests in 10.00s, 6.44MB read
Requests/sec:   4358.66
Transfer/sec:    659.76KB

Change-Id: I191450df06dca264b9ff541ba72da6179be7d21e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>