util: Count also the number of runs for every timer
util.Clock() keeps now the number of runs for every timer. Once we have
clock per connection, this will report the number of read and write
requests for the entire image.

Here is an example clock (from the clock benchmark) - without this change:

    [connection=0.468424, request=0.384151, read=0.074610, write=0.074435]

With this change:

    [connection=0.340290/1, request=0.271159/50000, read=0.049629/50000,
    write=0.050324/50000]

Today we have to calculate the total number of requests by grepping the
logs for WRITE/ZERO lines per ticket. With this change, we can switch
the WRITE and ZERO logs to debug level and fix the issue of huge logs
and loosing logs by rotation.

Note: to keep the format cleaner, we don't mark now unstopped timers
with "*".

Change-Id: I150af9e4d7b1e4e613b01e8769c91a39ec8ebfa8
Signed-off-by: Nir Soffer <nsoffer@redhat.com>