tests: Fix flaky tests
Commit 8dcb7235177f (daemon: Extend ticket on every request) added some
flaky tests assuming that one client completed reading response data the
server thread has closed the operation and extended the ticket. This
usually works but once in 10 tests it fails in the CI because the server
thread did not run yet.

Add 100 milliseconds sleep after reading the response, to make sure the
server thread run and close the operation.

Change-Id: I53e1ed9600e2f5686d8031c83ecafe8a1ed29b88
Signed-off-by: Nir Soffer <nsoffer@redhat.com>