http: Use utf-8 header encoding
We pass utf-8 filename in content-disposition header, so we cannot
latin1 encoding as used by python builtin web server.

Note that using utf-8 in headers is not compliant with HTTP RFCs, but we
used it successfully for several years. The correct way to do this is
described in https://tools.ietf.org/html/rfc5987. We may like to improve
later if we have time.

Change-Id: I8e8dcb8f1c27eb23fbf7b7c3ce9e86f8f6dad599
Signed-off-by: Nir Soffer <nsoffer@redhat.com>