web: Use plain text errors
We were using JSON errors including the error code, name, explanation,
and detail. The error code and name duplicate the response status code
and status text. The explanation is long and mostly unhelpful text
defined by webob http exception. The only useful data in that fancy JSON
is the detail, the text added by the code raising the error.

Change the error to plain text message with the detail. Update some
tests that expected a JSON error.

Since error format was not documented, this should not break clients.

Change-Id: I777f033aa62b89cbc6397b8b394788524c819f86
Signed-off-by: Nir Soffer <nsoffer@redhat.com>