file: Support unaligned writes
The file backend supports now unaligned writes using direct I/O.

file.BaseIO.write() has now two code paths:
- slow path - doing read-modify-write of a single block.
- fast path - doing fast write aligned to block size.

The slow path is used when writing to unaligned offset, or the buffer
size is smaller then a block size. In this case we read entire block
from storage, modify it, and write it back. This is more correct than
the previous method of disabling direct I/O when doing unaligned write.

The fast path is used when offset is aligned. If the buffer is not
aligned to block size, we write the aligned part of the buffer.

Both case may do a short write, so the caller must check the number of
bytes written and call again with the rest of the buffer. ops.Receive
already handle this case.

Writing to the last block of the file will pad the file with zeros if
the file size was not aligned to block size. A new tests was added to
ops tests to verify this behavior, and daemon tests were changed to
ignore padding.

Change-Id: Id18156457a478c61e1b76bd857bf7bf5de329db7
Signed-off-by: Nir Soffer <nsoffer@redhat.com>