util: Extract util.open() helper
Extract util.open() helper to open a io.FileIO() with direct I/O. This
helper will be used in ioutil tests instead of abusing file.open.

While moving, remove the unused direct=True option for the file backend.
We don't have code that need this option, and we removed the code to
open a file without direct I/O and enable it later.

Change-Id: I466af7dedb1c044ed5f595571d4e15d7c5217c45
Signed-off-by: Nir Soffer <nsoffer@redhat.com>