conftest: Return string instead of py.path.local
pytest tmpdir.join() returns a py.path.local object, which does not
support truncate, and has unfamiliar interface. We typically convert it
to a string in the test and use regular io.open() for modifying the file.
Return a string in the fixture to avoid the boilerplate in the tests.

Change-Id: Ic55ff5e6e2d042efeacfdf15ad436edb04ef3f3e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>