daemon: Cache backend per connection
Instead of opening and closing the backend on every request, open a
backend on the first request, and cache it in the connection context.
The backend is closed when a connection is closed.

There are several advantages with this:

- Correctness - fsync() on the same file descriptor avoid possible
  issues with fsync() error on multiple file descriptors.

- Performance - avoiding the syscalls for opening a file backend,
  querying the file properties, closing it, and flushing pending data to
  remote server with NFS. On NBD, avoiding slow handshake for every
  request.

This change also open the door for future performance improvements like
asynchronous and concurrent I/O.

Change-Id: I8bcada5f7c9f027fab183954e71d59cdd819c8f2
Signed-off-by: Nir Soffer <nsoffer@redhat.com>