tests: Add qemu_nbd.open helper
qemu_nbd.open() runs qemu-nbd with a given image in any format and
return a connected nbd client that can be used to read and write to the
image.

Example usage:

    subprocess.check_call(
        ["qemu-img", "create", "-f", "qcow2", "disk.qcow2", "1g"])

    with qemu_nbd.open("disk.qcow2", "qcow2") as c:
        c.write(104857600, b"it works")
        c.flush()

    # Backup disk...

    with qemu_nbd.open("backup.qcow2", "qcow2", read_only=True) as b:
        assert b.read(104857600, 8) == b"it works"

This helper will be used to populate disks with data before tests, or to
verify disk data after tests.

Change-Id: I789f6eec1edde74e08d032db179a2b41587f402e
Signed-off-by: Nir Soffer <nsoffer@redhat.com>