tests: Add nbd_server fixture
To make it easy to test the nbd backend, add nbd_server fixture.
The server creates a temporary file and socket for you, but is not
started so the test can configure the server as needed. The test must
start the server, but the test framework will stop the server when the
test ends.

Here is example usage, creating a read only server:

    def test_thing(nbd_server):
        nbd_server.read_only = True
        nbd_server.start()
        # Test code...

Change-Id: I517007fd107922cb8a5a681740b7f49e4ece9dd5
Signed-off-by: Nir Soffer <nsoffer@redhat.com>