nbd: Fix Client.close()
When closing a Client we added a new instance variable self._lock
because of unfortunate typo instead of setting self._sock to None.
However all this effort is not needed; if we just close the socket, any
operation on the socket will fail with:

    socket.error(EBADF, "Bad file descriptor")

Which is nicer than the AttributeError we get if we replace the socket
with None as we tried to do.

Change-Id: I603b9e2ef029cfaff7132458a602803ff0cca75b
Signed-off-by: Nir Soffer <nsoffer@redhat.com>