ops: Simplify Send loop
Move chunk writing into _next_chunk() and rename it to _send_chunk().
This remove duplicate code for writing and simplify the code.

This change also improve progress reporting; previously we increased the
done counter before writing the data to the client, and now we increase
only after the write was completed.

Change-Id: Ie3c780b7484ab9051b1d42b4443a55db202f60df
Signed-off-by: Nir Soffer <nsoffer@redhat.com>