packaging: setup: proxy: Use apache pki instead of own
Use apache httpd key/cert/ca_cert instead of engine-ca-generated ones.
This should make it easier to use:
- If using engine-ca for apache, user most likely already approved it
in the browser when logging in, so will not have to handle again for
imageio.
- If using 3rd party CA for apache, CA cert likely already added to the
browser.

In a normal upgrade, where the user didn't touch the conf manually,
notify the user and update the proxy conf.

In an upgrade where the user changed the conf manually after setup,
create a new file with a different name and notify the user, suggesting
to check the diff and update.

In an upgrade where the user changed the conf manually after setup, but
changed it to have the content we want it to have with current patch,
only log, but still "update" the file, so that a following setup or
especially cleanup will consider the file "ok" (not changed manually).

The only case that will be broken by current patch is if a user needs,
for some reason, two different certs for the two different https
services on two different ports, on the engine machine. In a discussion
in bugzilla it was decided that this is not an issue.

Bug-Url: https://bugzilla.redhat.com/1637809
Change-Id: I7bed070cfd8a84306e13a941815947e42c94bf3a
Signed-off-by: Yedidyah Bar David <didi@redhat.com>