backup: Workaround cleanup error with qemu-4.0.0-rc0
With qemu-4.0.0-rc0 we have a random failure during backup cleanp in the
CI:

    CommandFailed: Command {'arguments': {'name': 'sda'}, 'execute': 'nbd-server-remove'}
    failed: {u'desc': u"export 'sda' still in use", u'class': u'GenericError'}

I think the issue is shutting down and closing the NBD client in the
test process, and immediately tearing down the NBD server before qemu
detected that the NBD client disconnected.

Add a short sleep before tearing down the NBD server.

Change-Id: I52e5f9ebd6a6f7feb98e85455bc69ba31135dd0f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>