tests: Fix flaky download progress test
This test fails randomly with this error:

>       assert 0 < ticket.transferred() < size
E       AssertionError: assert 0 < 0

The issue is asserting about ticket status updated by another thread,
without a proper synchronization mechanism between the threads.

Add a short sleep after the client read from the socket, before checking
the ticket transferred size.

Since we added a new sleep, we are more likely to complete the download
on the server side, so increase the image size to 50 MiB to minimize the
chance of this event.

Change-Id: I57948071a759b921f6cb74b11cf3a79969323236
Signed-off-by: Nir Soffer <nsoffer@redhat.com>