testutil: Enhance wait_for_socket to allow no wait
Change the wait loop so we can call with timeout=0. This is useful for
testing the helper, or for checking that a socket was closed.

Improve the tests to validate the wait time, and use zero wait when we
can for faster tests.

Add debug log when starting the wait, and when ending show the wait time
in microseconds instead of milliseconds.

Change-Id: I73eb9f2dccdae3aac745462c662a6e7c88512efb
Signed-off-by: Nir Soffer <nsoffer@redhat.com>