nbd: Support NBD_OPT_GO information requests
Previously we sent empty information request list, but we may want to
send NBD_INFO_BLOCK_SIZE, and we may need to send other options to
implement NBD_CMD_BLOCK_STATUS.

Extract _format_go_option_data() for building the option data. This is
now a pure function that is easy to test and cannot be used in the wrong
way.

Merge _send_go_option() and _receive_go_reply() to
_negotiate_go_option() which is more clear, more useful, and cannot be
used in the wrong way.

Change-Id: Ibc4f0bea732d167e243c59fffbb87ff0a846c9ef
Signed-off-by: Nir Soffer <nsoffer@redhat.com>