nbd: Consume unknown reply data
NBD client MUST ignore information replies it does not understand, and
we do so, but we forgot to consume the reply data. This will lead to
failure when trying to process the next reply, which will fail the
negotiation.

Now we consume the unknown reply data, and log a warning including the
reply data, so we are more likely to fix the client to handle the reply.

Change-Id: I8f92ed98ea21ba02dd5a434d6aac532e889ac32c
Signed-off-by: Nir Soffer <nsoffer@redhat.com>