nbd: Introduce ReplyError exception
ReplyError is a request error reported when receiving a simple or
structured reply. In simple reply, we get only an error code, but in
structured reply we get also an error message suitable for displaying to
the user.

Previously we use to wrongly try to map the NBD error code using
NBD_REP_ERR constants, which are used only during negotiation. Now we
map the possible errors[1] correctly to system errno, and use os.strerror()
to get the error message.

[1] https://github.com/NetworkBlockDevice/nbd/blob/master/doc/proto.md#error-values

Change-Id: I0ce410adafd026786b85c8cfdd902556810cc6c0
Signed-off-by: Nir Soffer <nsoffer@redhat.com>