nbd: Fix NBD URL parsing
In the nbd tests, we tested string URLs like:

    "nbd:localhost:10900"

But in the application we get the URL from ticket JSON, so we get:

    u"nbd:localhost:10900"

While this is really the same thing in both python 2 and python 3, turns
out that socket.getaddrinfo() accepts port="10900", but not
port=u"10900", and we fail in runtime with:

    Traceback (most recent call last):
      ...
        sock = socket.create_connection(address)
      File "/usr/lib64/python2.7/socket.py", line 557, in create_connection
        for res in getaddrinfo(host, port, 0, SOCK_STREAM):
    error: getaddrinfo() argument 2 must be integer or string

This revealed that we had a bug in nbd._parse_url(), not converting the
port string to integer. This bug was hidden by socket.getaddrinfo()
accepting a string port.

Fixed by:

- Change nbd.TCPAddress() to validate the arguments, so now we fail
  early with ValueError if called with incorrect types.

- Change _parse_url() to convert the port to integer.

- Change the test to use Unicode URLs, testing real input.

Change-Id: I7335eb78aac36bb71f43577c58df282484fe565a
Signed-off-by: Nir Soffer <nsoffer@redhat.com>