http: Add failing tests for handling socket.error
If handler on the server side failed with socket.error, for example nbd
backend failing to connect to qemu because of bad port value:

    Traceback (most recent call last):
      ...
      File "/usr/lib64/python2.7/site-packages/ovirt_imageio_common/nbd.py", line 198, in open
        return Client(address, name)
      File "/usr/lib64/python2.7/site-packages/ovirt_imageio_common/nbd.py", line 268, in __init__
        self._sock = self._connect(address)
      File "/usr/lib64/python2.7/site-packages/ovirt_imageio_common/nbd.py", line 326, in _connect
        return self._create_tcp_connection(address)
      File "/usr/lib64/python2.7/site-packages/ovirt_imageio_common/nbd.py", line 339, in _create_tcp_connection
        sock = socket.create_connection(address)
      File "/usr/lib64/python2.7/socket.py", line 557, in create_connection
        for res in getaddrinfo(host, port, 0, SOCK_STREAM):
    error: getaddrinfo() argument 2 must be integer or string

http.Router treated this error as an error on the HTTP connection
socket, and raised without logging the error. This caused the connection
to be closed, and nothing was logged to server logs. The traceback
sneaked to stderr, where it was captured by systemd and logged to the
system journal.

We had a TODO to handle this case, but we never had code that trigger
this issue.

Add failing tests triggering this issue by raising a fake socket.error.

Change-Id: I9c167c6d38b57e747f00d5ff4fabd390dd60a6ef
Signed-off-by: Nir Soffer <nsoffer@redhat.com>