http: Fix unrelated socket.error handling
Fix error handling when request fail with a socket.error. Previously the
error was always handled as if the socket.error was related to the
HTTP connection socket, and the error as either logged using debug log,
or raised and logged to the system journal.

Now, if the error is related to the client connection and the connection
was lost, we close the connection and log a debug message. Otherwise we
log a traceback to the server log, and try to return a meaningful error
to the client.

Change-Id: I444e77fae0751f8f8af79998ac328f0c54c4d896
Signed-off-by: Nir Soffer <nsoffer@redhat.com>