file: Optimize trim at end of file
When we zero at the end of a sparse file, we use a combination of
truncate() and fallocate(). This seems to trigger buggy allocation in
NFS, leading to reporting wrong allocation during an upload.

Here is example upload flow showing this issue, uploading a
fully-allocated qcow2 image.

1. Writing qcow2 header

[images] [10.35.161.127] WRITE size=200704 offset=0 flush=False
[backends.file] STAT op=write size=200704 allocated=200704

2. Writing small zero area after qcow2 header

[images] [10.35.161.127] ZERO size=61440 offset=200704 flush=False
[backends.file] STAT op=trim size=262144 allocated=200704

Zero is implemented by:

    truncate(262144)
    fallocate(PUHCH_HOLE | KEEP_SIZE, 200704, 61440)

(This is required by zeroing a range starting before the end of the
 file, ending after the end of the file).

3. Writing the rest of the data (128.0625 MiB)

[images] [10.35.161.127] WRITE size=134283264 offset=262144 flush=False

[backends.file] STAT op=write size=8650752 allocated=8589312
[backends.file] STAT op=write size=17039360 allocated=16977920
[backends.file] STAT op=write size=25427968 allocated=25366528
...
[backends.file] STAT op=write size=117702656 allocated=117641216
[backends.file] STAT op=write size=126091264 allocated=126029824
[backends.file] STAT op=write size=134479872 allocated=134418432

Up to this point allocation looks normal...

[backends.file] STAT op=write size=134545408 allocated=201527296

After this write, file size is 134545408 (128.3125 MiB) but
allocation is 201527296 (192.191 MiB).

4. Flushing

[images] [10.35.161.127] FLUSH
[backends.file] STAT op=flush size=134545408 allocated=201527296

5. Closing

[http] CLOSE client=10.35.161.127 [connection=10.076871/1,
dispatch=9.998022/5, operation=9.977598/4, read=8.650615/18,
write=1.149925/18, zero=0.005621/1, flush=0.002144/1]
[backends.file] STAT op=close size=134545408 allocated=201527296

Allocation remains incorrect after closing the file. Few seconds later,
allocation go down to 128.3125 MiB but at this point engine already
stored the incorrect allocation in the db.

Change the way we zero to skip the fallocate() call in the special case
of zeroing at the end of the file.

Avoiding fallocate() save one syscall per zero operation. Since this is
the normal flow during upload, this safe this can also speed up zeroing
of sparse files.

Change-Id: Ib48e64295e77cb6197e10e2db6aca95f4338769e
Bug-Url: https://bugzilla.redhat.com/1730273
Signed-off-by: Nir Soffer <nsoffer@redhat.com>