tests: Use userstorage for file backend tests
Previously we use a temporary file, limiting the tests to file storage
and the underlying storage sector size. Replace the tempurl fixture with
user_url, backed by userstorage file backend with sector size of 512
bytes.

Since ext2 does not support fallocate(), the tests for preallocated
storage trigger the fallback to manual zeroing, which is not compatible
with python 3 yet. We don't have a good way to mark as xfail only the
ext2 preallocated tests, so all preallocated tests are mark as xfail on
python 3.

Change-Id: I238bbeb734160063bf217c9325adc8d50417b48e
Bug-Url: https://bugzilla.redhat.com/172980
Signed-off-by: Nir Soffer <nsoffer@redhat.com>