file: FileBackend 4k support
Detect underlying storage block size when opening a file backend,
instead of using hard-coded value. With this change unaligned I/O
succeeds.

On XFS reading from a hole succeeds even if the buffer is smaller than
the required alignment, so we detect the block size by reading one block
from storage and writing the block back to storage.

In the special case of empty or tiny files (size < block size), writing
one block with direct I/O will enlarge the file to size of one block. We
truncate the file back to the original size after detecting the size,
but if truncate will fail we may leave the file with modified size. Such
images are not expected so I think we are fine.

It is possible to detect block size on XFS filesystem using xfsctl[1],
avoiding the write to storage. However this will help only local XFS
filesystem which are not our main use case, and will not help to handle
other filesystems using the same optimization for reading holes.

[1] https://linux.die.net/man/3/xfsctl

Change-Id: Id3b894eeee686cb2289c77c4bc4addf9ce9c15e4
Bug-Url: https://bugzilla.redhat.com/172980
Signed-off-by: Nir Soffer <nsoffer@redhat.com>