backends: file: Fix detect block size for Gluster
When Gluster volume "performance.strict-o-direct" is "off", unaligned
write may succeed, because Gluster may buffer the write. The write will
may fail later when Gluster perform the actual write to storage, or when
issuing fsync().

It seems that this option is required for oVirt setup, so we may not
need to support Gluster storage with this option disabled, but since
this is not the default Gluster configuration, it is likely that we will
work with storage configured in this way.

Opening the file with O_SYNC avoids this issue, since this flag disable
the write-behind cache in Gluster[1]. Add it so imageio is more robust
with misconfigured Gluster volumes.

[1] https://github.com/gluster/glusterfs/blob/25cf0d24caa11bdfefe10c93541da07c635d4cfe/xlators/performance/write-behind/src/write-behind.c#L1866

Change-Id: I60d05da9acaad26f554b8015df8ee6e256d6c541
Signed-off-by: Nir Soffer <nsoffer@redhat.com>