automation: Add python 3 build for el8
Enable el8 for all builds, and add automation files for el8.

On Travis all tests pass, but on Jenkins check-patch.sh fail to create
the "file-512-xfs" backend:

    [userstorage] INFO    Creating filesystem /var/tmp/imageio-storage/file-512-xfs-mount
    mount: /var/tmp/imageio-storage/file-512-xfs-mount: wrong fs type,
    bad option, bad superblock on /dev/loop2, missing codepage or helper
    program, or other error.

Mark the "file-512-xfs" backend as optional, so we will continue the build
even if creating some storage fails. We need to make this more strict
later when we understand this failure.

To handle mount failures, require userstorage 0.4, supporting failures
when creating filesystems and mounting devices for optional storage
backend.

Change-Id: I9fcb429237a31df3b55b99a6d11938b469039fe5
Signed-off-by: Nir Soffer <nsoffer@redhat.com>