common: Move web module to proxy
The web module is used only by the proxy, so it should be in the proxy
package. This remove the python3-webob dependency which does not exist
in CentOS8.

To move the module, testutil.UbufferedStream test helper moved to
ovirt_imageio_common.util so the proxy can use it in the web module
tests.

Change-Id: I4e78c16e5dc7882c71274dd66c66bbdc74407057
Signed-off-by: Nir Soffer <nsoffer@redhat.com>