build: Support building packages for python 3
We want to build ovirt-imageio-common and ovirt-imageio-daemon both for
python 2 and python 3, and ovirt-imageio-proxy only for python 2.

The planned build matrix is:

daemon, common:
- rhel 7, fedora 29: python 2
- rhel 8, fedora 30: python 3

proxy:
- rhel 7, fedora 29: python 2

We are keeping the daemon python 2 builds to make it easier to test
fixes for 4.3. We will drop them when 4.4 is released.

Because we never build both python 2 and 3, we don't need to change the
packages names.

Add PYTHON_VERSION argument to the daemon and common makefiles to allow
building for both python 2 and python 3.

To build for python 3 run:

    make PYTHON_VERSION=3
    make PYTHON_VERSION=3 rpm

Change-Id: Ie6eddb72baf1a6eb275bf2d4cec3479551d6c2c7
Signed-off-by: Nir Soffer <nsoffer@redhat.com>