tox: Timeout tests after 30 seconds
Recently we have too many hang tests, maybe because the CI slaves are
overloaded with large number of concurrent builds. This requires
checking and aborting the jobs manually which is too time consuming.

Use pytest-timeout plugin[1] to timeout out hang tests, printing
backtrace of all threads to stderr, and killing the test process.

[1] https://pypi.org/project/pytest-timeout/

Change-Id: I76eb48c88be1a5492488d8f64fd2f1bd73c2ed3f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>