tests: Use random TCP port for qemu/qemu-nbd tests
We used a NBD default TCP port for some tests; this cause random hangs
in Jenkins, when another process is using this port. The test times out
with this error log:

    qemu-nbd: Failed to find an available port: Address already in use

Change the tests using TCP transport to use random port selected by the
kernel, which is unlikely to be used when qemu or qemu-nbd will try to
bind to the port.

Change-Id: I43feae1391fbf8493c853e0c2b4342432d0bcf94
Signed-off-by: Nir Soffer <nsoffer@redhat.com>