tox: Fix daemon coverage report
In the past we did not have enough tests for the common package, so we
added it to the daemon tests to see which code is covered. Since some
common modules are not used by the daemon this reduced the daemon
coverage from 89% to 63%.

Remove the ovirt_imageio_common package so we get correct coverage for
the daemon package.

Change-Id: Idb435da341d786021a9532782e86d26d96fb87fc
Signed-off-by: Nir Soffer <nsoffer@redhat.com>