compat: Use subprocess32 on python 2
On Jenkins we still have hang tests after fixing the TCP port issue. In
a recent timeout we see that the test is hang waiting for qemu-nbd:

  File "/home/jenkins/workspace/ovirt-imageio_standard-check-patch/ovirt-imageio/common/test/qemu_nbd.py", line 77, in stop
    self.proc.wait()
  File "/usr/lib64/python3.7/subprocess.py", line 990, in wait
    return self._wait(timeout=timeout)
  File "/usr/lib64/python3.7/subprocess.py", line 1624, in _wait
    (pid, sts) = self._try_wait(0)
  File "/usr/lib64/python3.7/subprocess.py", line 1582, in _try_wait
    (pid, sts) = os.waitpid(self.pid, wait_flags)

Use subprocess32 on python 2, supporting timeout in Popen.wait(). Wait
one second after terminating qemu or qemu-nbd, and if the process did
not terminate kill it.

Change-Id: Ia53260247df35e9b3d5071bdb255f27a815670a4
Signed-off-by: Nir Soffer <nsoffer@redhat.com>