daemon: Fix delay when stopping service
Looks like python 3 changed the behaviour of time.sleep(), and now it
is not interrupted by signals. When we stop the server we can have up to
30 seconds delay before the server notice that termination signal was
received. This delay slows down host reboot for no reason.

Replace time.sleep(30) with signal.pause(), returning immediately when
receiving a signal. Now we stop the http servers immediately when
receiving a termination signal.

Change-Id: I5d11749372449f585d803da473b35e6b8a6fe7e9
Signed-off-by: Nir Soffer <nsoffer@redhat.com>