test: Test NBD extents also with qcow2 format
Extend the tests to use both raw and qcow2 images.

Create the images with qemu-img, and use the NBD client to populate the
image. This allows testing with both raw and qcow2 format.

For empty images, raw and qcow2 report different extents so the tests
were separated.

Empty raw images are not created with qemu-img, since we currently have
inconsistent allocation in Fedora and CentOS 8.0. We can use qemu-img
once we require CentOS 8.1 AV.

For qcow2 images we also have inconsistent allocation in CentOS and
Fedora; write full cluster to avoid that.

Change-Id: I5594a364f7e58d4fc41a77883e925f32509e07c4
Signed-off-by: Nir Soffer <nsoffer@redhat.com>