tests: Extract helper function to read all extents
I want to add a new tests for handling many extents, and we will need
the same extent loop and merge for this test. Extract the code to a
helper function.

I wonder if we should integrate this function in nbd.Client.extents().
It will be more efficient to merge extent as they are read from the
socket, and it will make the tests and the nbd backend simpler.

Change-Id: I9865e8dd7208934a863871236a94ff81da68417f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>