nbd: Improve memory handling when reading extents
Server can return up to 1048576 extents per 4GiB in raw format, but we
expect to get less then 1000 extents per call. It does not make sense to
allocate a buffer that will hold the entire payload.

Change reading block status payload to allocate a small buffer and read
all extents in a loop, reading up to 1024 extents per iteration. We
expect to do one iteration in most cases.

Add a test generating and reading 2000 tiny extents. Since qcow2 minimum
extent size is a cluster (64 KiB), test this only with raw format, when
extent size is the file system block size (most likely 4 KiB).

We still do not limit the number or extents. The NBD spec is not clear
about this so this require some heuristics.

Change-Id: Iceb0eb12dcf96aca9ff4c79f7a945cce497d1b23
Signed-off-by: Nir Soffer <nsoffer@redhat.com>