nbd: Fix handling of incomplete structured reply error
When handling structured reply error chunk (NBD_REPLY_TYPE_ERROR) raise
ReplyError only if this was the last chunk (REPLY_FLAG_DONE is set). In
this case if the client can recover from the error, it can send the next
request.

If this is not the last chuck raise ProtocolError, failing the entire
connection. It may be possible to handle this case by consuming the next
chunks before failing but I don't think it is worth the effort, since we
never seen qemu return such error.

Improve the documentation of RequestError subclasses about this, and
change handling of the error to make it clear that handling error chunk
always raises.

Change-Id: I74f90680d498f61505f6e330807dd6a7c3eae6c7
Signed-off-by: Nir Soffer <nsoffer@redhat.com>