nbd: Handle NBD_REPLY_TYPE_ERROR_OFFSET for NBD_CMD_BLOCK_STATUS
According to NBD spec, a server may return one or more
NBD_REPLY_TYPE_ERROR_OFFSET chunks in reply to NBD_CMD_BLOCK_STATUS.

To handle this we must keep the errors and continue until we get the
last chunk of the structure reply.

In this case we raise a RequestError with the list of received errors.
If the client can recover from the error it can continue normally and
send new requests.

Payload length validation moved down to _receive_block_status_payload
where it belong.

Not tested since qemu does not send this error chunk.

Change-Id: I7e7357fba67d95593c47bd63bd3993ef10bab989
Signed-off-by: Nir Soffer <nsoffer@redhat.com>