nbd: Move only_structured to Command
This flag was set only when calling _recv_reply() with a Read command,
so it belongs to commands classes.

BlockStatus command supports only structured replies, so it overrides
this flag to True.

Read command supports enforcing structure replies if the server support
structured replies, so it accepts a only_structured argument.

Other commands support both simple and structured replies. They start
with only_structured=False. We set the value to True if a command
received the first structured reply chunk. From this point simple reply
is forbidden.

Change-Id: I106150eccada961079c907980b9f43eecbfb94c0
Signed-off-by: Nir Soffer <nsoffer@redhat.com>