nbd: Move buf into Commands
Read is the only command using a buffer to receive payload from the
server. We used to pass a None default buffer around, which is error
prone and confusing.

Move buf to Command.buf, defaulting to None. Read command adds a
required buf argument in __init__(), so it is easy to create it
correctly.

Remove buf from methods signature, and access cmd.buf instead. For
consistency, also remove the offset argument since it can be accessed
via cmd.offset.

Change-Id: Iecc2549a899715782219288d513c3fbeb20c07bf
Signed-off-by: Nir Soffer <nsoffer@redhat.com>