tests: Fix flaky PUT tests
If a client sends an invalid PUT request with a payload, we respond
immediately with 40X error without reading the payload, so we have to
close the connection.

Tests simulating this fail randomly on Jenkins with:

    def send(self, data):
        ...
        try:
            self.sock.sendall(data)
            BrokenPipeError: [Errno 32] Broken pipe

In real world this should not be an issue, since client should start
their session with OPTION request which does not have this issue.

Make the tests more reliable, by making the http client more careful; if
sending the body fail with EPIPE, try to return the response from the
server anyway. If the server did send a response the tests should
succeed.

Same method is used successfully in imageio client and virt-v2v.

Change-Id: Id9a33f7f3ae6cc5f84eec6362abb1a9ffa9cb5ef
Signed-off-by: Nir Soffer <nsoffer@redhat.com>