examples: Multi-threaded copy
Our nbd client does not support multiple in-flight requests, but when
using 2 clients in the copy example, we can boost performance by running
each client in its own thread.

Initial testing on my laptop shows 40% improvement, but the results are
not stable enough. We need to test this in a real environment.

Before:

$ hyperfine --warmup 1 './nbd-client -s copy nbd:unix:/tmp/src.sock nbd:unix:/tmp/dst.sock'
  Time (mean ± σ):      2.698 s ±  0.078 s    [User: 70.9 ms, System: 344.4 ms]
  Range (min … max):    2.623 s …  2.879 s    10 runs

After:

$ hyperfine --warmup 1 './nbd-client -s copy nbd:unix:/tmp/src.sock nbd:unix:/tmp/dst.sock'
  Time (mean ± σ):      1.885 s ±  0.079 s    [User: 94.7 ms, System: 386.7 ms]
  Range (min … max):    1.803 s …  2.087 s    10 runs

For reference, same operation using qemu-img convert:

$ hyperfine --warmup 1 'qemu-img convert -n -f raw -O raw nbd:unix:/tmp/src.sock nbd:unix:/tmp/dst.sock'
  Time (mean ± σ):      2.277 s ±  0.474 s    [User: 132.1 ms, System: 457.5 ms]
  Range (min … max):    2.027 s …  3.424 s    10 runs

Change-Id: Id1691ed8301e872b330b02a0295996a927da826b
Signed-off-by: Nir Soffer <nsoffer@redhat.com>